move ts join to other class
